cleaning API.

+++ include/xorfilter.h
@@ -122,7 +122,7 @@ static inline bool xor8_allocate(uint32_t size, xor8_t *filter) {
 
 // allocate enough capacity for a set containing up to 'size' elements
 // caller is responsible to call xor8_free(filter)
-static inline bool xor16_allocate(size_t size, xor16_t *filter) {
+static inline bool xor16_allocate(uint32_t size, xor16_t *filter) {
   size_t capacity = 32 + 1.23 * size;
   filter->blockLength = capacity / 3;
   capacity = capacity / 3 * 3;
@@ -161,7 +161,7 @@ static inline void xor16_free(xor16_t *filter) {
 
 struct xor_xorset_s {
   uint64_t xormask;
-  uint64_t count;
+  uint32_t count;
 };
 
 typedef struct xor_xorset_s xor_xorset_t;
@@ -263,7 +263,7 @@ static inline xor_h0h1h2_t xor16_get_just_h0_h1_h2(uint64_t hash,
 
 struct xor_keyindex_s {
   uint64_t hash;
-  uint64_t index;
+  uint32_t index;
 };
 
 typedef struct xor_keyindex_s xor_keyindex_t;
@@ -455,7 +455,7 @@ static inline uint32_t xor_flushone_decrement_buffer(xor_setbuffer_t *buffer,
 // size is the number of keys
 // the caller is responsable for calling xor8_allocate(size,filter) before
 //
-bool xor8_buffered_populate(const uint64_t *keys, size_t size, xor8_t *filter) {
+bool xor8_buffered_populate(const uint64_t *keys, uint32_t size, xor8_t *filter) {
   uint64_t rng_counter = 1;
   filter->seed = xor_rng_splitmix64(&rng_counter);
   size_t arrayLength = filter->blockLength * 3; // size of the backing array
@@ -811,7 +811,7 @@ bool xor8_populate(const uint64_t *keys, size_t size, xor8_t *filter) {
 // size is the number of keys
 // the caller is responsable for calling xor8_allocate(size,filter) before
 //
-bool xor16_buffered_populate(const uint64_t *keys, size_t size, xor16_t *filter) {
+bool xor16_buffered_populate(const uint64_t *keys, uint32_t size, xor16_t *filter) {
   uint64_t rng_counter = 1;
   filter->seed = xor_rng_splitmix64(&rng_counter);
   size_t arrayLength = filter->blockLength * 3; // size of the backing array
@@ -1005,7 +1005,7 @@ bool xor16_buffered_populate(const uint64_t *keys, size_t size, xor16_t *filter)
 // size is the number of keys
 // the caller is responsable for calling xor8_allocate(size,filter) before
 //
-bool xor16_populate(const uint64_t *keys, size_t size, xor16_t *filter) {
+bool xor16_populate(const uint64_t *keys, uint32_t size, xor16_t *filter) {
   uint64_t rng_counter = 1;
   filter->seed = xor_rng_splitmix64(&rng_counter);
   size_t arrayLength = filter->blockLength * 3; // size of the backing array
